Parse printed  mesages with normalized line endings
Parse pretty printed messages while normalizing line endings to it is
agnostic to whatever OS is the user using.

Not using VSCode api for determining the line ending since the function
uses a string as parameter and there is not Window or Document object
involved in current functionality. Just good old search and replace to
assume consistent line endings.